make-driven build of aws-sigv4-proxy
This facilitates building the proxy image in an environment where
multi-stage builds aren't available to docker.  It uses make with
bash, but is fairly light-touch in the bashisms so as to be
reasonably portable.

Just do `make` and you'll get a tagged aws-sigv4-proxy:latest docker
image that runs the proxy.

This is initialized to work of brianbianco's aws-sigv4-proxy fork
to get the header strip feature.